Changed compiler settings to make unused imports and variables and unnecessary casts to be errors.

+++ trunk/source/org.eclipse.ve/plugins/org.eclipse.ve.swt/vm_swt/org/eclipse/ve/internal/swt/targetvm/ControlManager.java
@@ -278,7 +278,7 @@ public abstract class ControlManager implements ICallback , ControlListener {
 							});
 							// Send back all of the image data
 							for (int i = 0; i < data.length; i++) {
-								outputStream.write((int) data[i]);
+								outputStream.write(data[i]);
 							}
 
 							outputStream.close();
